refactor(live): consolidate location functionality to single controller
- Remove HomeController and its test file (no longer needed)
- Consolidate location update functionality into LocationController
- Replace generic POST /locations with domain-specific PUT endpoint
- Update routes to single location endpoint: PUT /skatingEvents/:id/skaters/:id
- Update tests to focus on skatemap domain functionality
- All tests pass (2/2 successful)

BREAKING CHANGE: Removes GET / home page and POST /locations endpoints
