+++ src/main/resources/routes
@@ -3,7 +3,5 @@
 # https://www.playframework.com/documentation/latest/ScalaRouting
 # ~~~~
 
-# An example controller showing a sample home page
-GET     /                           controllers.HomeController.index()
-POST    /locations                  controllers.LocationController.update()
-PUT     /skatingEvents/:skatingEventId/skaters/:skaterId    controllers.HomeController.updateLocation(skatingEventId, skaterId)
+# Location update endpoint for skating events
+PUT     /skatingEvents/:skatingEventId/skaters/:skaterId    controllers.LocationController.updateLocation(skatingEventId, skaterId)
