--- src/main/scala/controllers/HomeController.scala
@@ -1,30 +1,0 @@
-package controllers
-
-import javax.inject._
-import play.api._
-import play.api.mvc._
-
-/**
- * This controller creates an `Action` to handle HTTP requests to the
- * application's home page.
- */
-@Singleton
-class HomeController @Inject()(val controllerComponents: ControllerComponents) extends BaseController {
-
-  /**
-   * Create an Action to render an HTML page.
-   *
-   * The configuration in the `routes` file means that this method
-   * will be called when the application receives a `GET` request with
-   * a path of `/`.
-   */
-  def index(): Action[AnyContent] = Action { implicit request: Request[AnyContent] =>
-    Ok
-  }
-
-  def updateLocation(skatingEventId: String, skaterId: String): Action[AnyContent] = {
-    Action { implicit request: Request[AnyContent] =>
-      Ok
-    }
-  }
-}
