+++ src/main/scala/controllers/LocationController.scala
@@ -8,8 +8,9 @@ import javax.inject.{Inject, Singleton}
 @Singleton
 class LocationController @Inject()(val controllerComponents: ControllerComponents) extends BaseController {
 
-  def update(): Action[AnyContent] = Action {
-    implicit request: Request[AnyContent] =>
+  def updateLocation(skatingEventId: String, skaterId: String): Action[AnyContent] = {
+    Action { implicit request: Request[AnyContent] =>
       Ok
+    }
   }
 }
